feat: Setting up Model for commit history
